requirements.txt
streamlit
PyPDF2
anthropic
python-dotenv
